chore(core): move tests within tests folder
